Because the rear and front values keep moving, the queue is full only when the rear and the front are side by side. So in my array of size 5, for example, when the rear value is 4 and the front value is 0, then the queue is full.
rear+1 is the position right next
to the current rear. (rear+1)%n will always return that next position (rear+1) if the value of (rear+1)
is not greater than or equal to five(the size of my array). if (rear+1)%n == front is basically asking "is the
next position from the end of the queue(rear) the beginning of the queue(front)"? And if
that is true, the queue must be full.